Move iterative length event and tasks to separate files
