+++ src/core/operator/event/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(EXTENSION_SOURCES
+        ${CMAKE_CURRENT_SOURCE_DIR}/iterative_length_event.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/shortest_path_event.cpp
         ${EXTENSION_SOURCES}
         PARENT_SCOPE
