+++ src/core/operator/physical_path_finding_operator.cpp
@@ -10,12 +10,12 @@
 #include "duckdb/parallel/event.hpp"
 #include "duckdb/parallel/meta_pipeline.hpp"
 #include "duckdb/parallel/thread_context.hpp"
+#include "duckpgq/core/utils/duckpgq_barrier.hpp"
 #include <algorithm>
 #include <duckpgq/core/operator/event/shortest_path_event.hpp>
 #include <thread>
-#include <utility>
 
-#include "duckpgq/core/utils/duckpgq_barrier.hpp"
+#include <duckpgq/core/operator/event/iterative_length_event.hpp>
 
 namespace duckpgq {
 
@@ -267,141 +267,6 @@ PhysicalPathFinding::Combine(ExecutionContext &context,
 // Finalize
 //===--------------------------------------------------------------------===//
 
-class PhysicalIterativeTask : public ExecutorTask {
-public:
-  PhysicalIterativeTask(shared_ptr<Event> event_p, ClientContext &context,
-                        PathFindingGlobalState &state, idx_t worker_id)
-      : ExecutorTask(context, std::move(event_p)), context(context),
-        state(state), worker_id(worker_id) {}
-
-  TaskExecutionResult ExecuteTask(TaskExecutionMode mode) override {
-    auto &bfs_state = state.global_bfs_state;
-    auto &change = bfs_state->change;
-    auto &barrier = bfs_state->barrier;
-
-    auto bound = bfs_state->BoundaryCalculation(worker_id);
-    left = bound.first;
-    right = bound.second;
-
-    do {
-      bfs_state->InitTask(worker_id);
-
-      IterativeLength();
-
-      barrier.Wait();
-
-      if (worker_id == 0) {
-        ReachDetect();
-      }
-      barrier.Wait();
-    } while (change);
-
-    if (worker_id == 0) {
-      UnReachableSet();
-    }
-
-    event->FinishTask();
-    return TaskExecutionResult::TASK_FINISHED;
-  }
-
-private:
-  void IterativeLength() {
-    auto &bfs_state = state.global_bfs_state;
-    auto &seen = bfs_state->seen;
-    auto &visit = bfs_state->iter & 1 ? bfs_state->visit1 : bfs_state->visit2;
-    auto &next = bfs_state->iter & 1 ? bfs_state->visit2 : bfs_state->visit1;
-    auto &barrier = bfs_state->barrier;
-    int64_t *v = (int64_t *)state.global_csr->v;
-    vector<int64_t> &e = state.global_csr->e;
-    auto &lane_to_num = bfs_state->lane_to_num;
-    auto &change = bfs_state->change;
-
-    // clear next before each iteration
-    for (auto i = left; i < right; i++) {
-      next[i] = 0;
-    }
-
-    barrier.Wait();
-
-    while (true) {
-      auto task = bfs_state->FetchTask(worker_id);
-      if (task.first == task.second) {
-        break;
-      }
-      auto start = task.first;
-      auto end = task.second;
-
-      for (auto i = start; i < end; i++) {
-        if (visit[i].any()) {
-          for (auto offset = v[i]; offset < v[i + 1]; offset++) {
-            auto n = e[offset];
-            std::lock_guard<std::mutex> lock(bfs_state->element_locks[n]);
-            next[n] |= visit[i];
-          }
-        }
-      }
-    }
-
-    change = false;
-    barrier.Wait();
-
-    for (auto i = left; i < right; i++) {
-      if (next[i].any()) {
-        next[i] &= ~seen[i];
-        seen[i] |= next[i];
-        change |= next[i].any();
-      }
-    }
-  }
-
-  void ReachDetect() {
-    auto &bfs_state = state.global_bfs_state;
-    auto result_data = FlatVector::GetData<int64_t>(bfs_state->result.data[0]);
-
-    // detect lanes that finished
-    for (int64_t lane = 0; lane < LANE_LIMIT; lane++) {
-      int64_t search_num = bfs_state->lane_to_num[lane];
-      if (search_num >= 0) { // active lane
-        int64_t dst_pos = bfs_state->vdata_dst.sel->get_index(search_num);
-        if (bfs_state->seen[bfs_state->dst[dst_pos]][lane]) {
-          result_data[search_num] =
-              bfs_state->iter; /* found at iter => iter = path length */
-          bfs_state->lane_to_num[lane] = -1; // mark inactive
-          bfs_state->active--;
-        }
-      }
-    }
-    if (bfs_state->active == 0) {
-      bfs_state->change = false;
-    }
-    // into the next iteration
-    bfs_state->iter++;
-  }
-
-  void UnReachableSet() {
-    auto &bfs_state = state.global_bfs_state;
-    auto result_data = FlatVector::GetData<int64_t>(bfs_state->result.data[0]);
-    auto &result_validity = FlatVector::Validity(bfs_state->result.data[0]);
-
-    for (int64_t lane = 0; lane < LANE_LIMIT; lane++) {
-      int64_t search_num = bfs_state->lane_to_num[lane];
-      if (search_num >= 0) { // active lane
-        result_validity.SetInvalid(search_num);
-        result_data[search_num] = (int64_t)-1; /* no path */
-        bfs_state->lane_to_num[lane] = -1;     // mark inactive
-      }
-    }
-  }
-
-private:
-  ClientContext &context;
-  PathFindingGlobalState &state;
-  // [left, right)
-  idx_t left;
-  idx_t right;
-  idx_t worker_id;
-};
-
 class PhysicalCSREdgeCreationTask : public ExecutorTask {
 public:
   PhysicalCSREdgeCreationTask(shared_ptr<Event> event_p, ClientContext &context,
@@ -478,35 +343,7 @@ public:
   }
 };
 
-class ParallelIterativeEvent : public BasePipelineEvent {
-public:
-  ParallelIterativeEvent(PathFindingGlobalState &gstate_p, Pipeline &pipeline_p)
-      : BasePipelineEvent(pipeline_p), gstate(gstate_p) {}
 
-  PathFindingGlobalState &gstate;
-
-public:
-  void Schedule() override {
-    auto &bfs_state = gstate.global_bfs_state;
-    auto &context = pipeline->GetClientContext();
-
-    vector<shared_ptr<Task>> bfs_tasks;
-    for (idx_t tnum = 0; tnum < bfs_state->num_threads; tnum++) {
-      bfs_tasks.push_back(make_uniq<PhysicalIterativeTask>(
-          shared_from_this(), context, gstate, tnum));
-    }
-    SetTasks(std::move(bfs_tasks));
-  }
-
-  void FinishEvent() override {
-    auto &bfs_state = gstate.global_bfs_state;
-
-    // if remaining pairs, schedule the BFS for the next batch
-    if (bfs_state->started_searches < gstate.global_tasks->Count()) {
-      PhysicalPathFinding::ScheduleBFSEvent(*pipeline, *this, gstate);
-    }
-  }
-};
 
 SinkFinalizeType
 PhysicalPathFinding::Finalize(Pipeline &pipeline, Event &event,
